BAC-5-3: other parts of refactor of project by FSD
