+++ src/shared/ui/Inputs/InputSplitter/index.ts
@@ -1,0 +1,1 @@
+export * from './ui';
