+++ src/shared/ui/Inputs/InputSplitter/ui/constants.ts
@@ -1,0 +1,17 @@
+export const LABEL = 'Enter words';
+export const PLACEHOLDER = 'Enter words to split';
+export const ROW = 1;
+
+export const BUTTON_REFRESH_CONFIG = {
+  icon: 'system_update_alt',
+  color: 'primary',
+  outline: true,
+  flat: true
+};
+
+export const BASE_INPUT_CONFIG = {
+  label: LABEL,
+  placeholder: PLACEHOLDER,
+  rows: ROW,
+  autogrow: true
+};
