+++ src/shared/ui/Inputs/InputSplitter/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as InputSplitter } from './InputSplitter.vue';
+export * from './constants';
