+++ src/shared/ui/Inputs/index.ts
@@ -1,0 +1,2 @@
+export * from './InputSplitter';
+export * from './InputSeparatedSymbol';
