+++ src/shared/ui/Words/WordListForm/ui/WordListForm.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import { defineProps, computed } from 'vue';
-  import { generateId } from 'src/shared/lib/utils/id-generator';
+  import { generateId } from '@/shared/lib/utils/id-generator';
   import { TABLE_CONFIG } from './constants';
   import { WordEntity } from '@/features/words/types/word';
