+++ src/shared/ui/Words/WordsItem/ui/WordsItem.vue
