+++ src/shared/ui/Words/WordsItem/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as WordsItem } from './WordsItem.vue';
