+++ src/shared/ui/Words/WordsList/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as WordsList } from './WordsList.vue';
