+++ src/shared/ui/Words/index.ts
@@ -1,0 +1,3 @@
+export * from './WordListForm';
+export * from './WordListForm';
+export * from './WordsList';
