+++ src/widgets/Words/ImportWords/ImportTextField/ui/ImportTextField.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
   import { ref } from 'vue';
   import { VModal } from 'src/shared/ui/VModal';
-  import InputSplitter from 'src/features/words/components/InputSplitter.vue';
+  import { InputSplitter } from 'src/shared/ui/Inputs';
+  import { WordListForm } from 'src/shared/ui/Words/WordListForm';
   import { WordEntity } from 'src/features/words/types/word';
   import { generateId } from 'src/shared/lib/utils/id-generator';
   import { BUTTON_CONFIG, DEFAULT_SEPARATED_VALUES, MODAL } from './constants';
-  import { WordListForm } from 'src/shared/ui/WordListForm';
   import {
     InputSeparatedSymbol,
     TITLE_SEPARATDE_ON_WORD_ITEMS,
