+++ src/widgets/Words/WordsForm/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as WordsForm } from './WordsForm.vue';
