+++ src/widgets/Words/types.ts
